+++ app/scaffold-config/baseConfig.js
@@ -1,6 +1,6 @@
-/* global AstroNative*/
-define([],
-function() {
+define([], function() {
+/* eslint-enable max-statements */
+
     // Line altered by generator. If changing please change generator as well!!
     var iosUsingTabLayout = false;
 
@@ -9,8 +9,12 @@ function() {
         secondaryColor: '#007ba7',
         whiteColor: '#ffffff'
     };
+
+    var loaderColor = colors.primaryColor;
+
     return {
         colors: colors,
-        iosUsingTabLayout: iosUsingTabLayout
+        iosUsingTabLayout: iosUsingTabLayout,
+        loaderColor: loaderColor
     };
 });